feat(react-router-aws): add build utilities and update adapter configuration
