+++ packages/react-router-aws/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@nucel.cloud/react-router-aws",
-  "version": "0.1.2",
+  "version": "0.1.3",
   "description": "React Router v7 adapter and Pulumi components for deploying to AWS Lambda, CloudFront, and S3",
   "type": "module",
   "main": "./dist/index.js",
